+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.7
 * Fixed typos by Morgan (https://github.com/rmtmckenzie) and Nico Mexis (https://github.com/ThexXTURBOXx)
 * Updated dependencies by Nico Mexis (https://github.com/ThexXTURBOXx)
+* Added `reportOccurrenceTimeout` in CatcherOptions to prevent reporting duplicated errors.
 
 ## 0.6.6
 * Added copyWith method in LocalizationOptions.